requirement.txt
numpy
pytorch==1.7.0
opencv-python
pyyaml
yacs
tqdm
colorama
matplotlib
cython
tensorboardX
